chore: organize authentication structure and improve formatting
