--- src/Ambev.DeveloperEvaluation.Application/Auth/AuthenticateUser/AuthenticateUserValidator.cs
@@ -1,18 +1,0 @@
-using FluentValidation;
-
-namespace Ambev.DeveloperEvaluation.Application.Auth.AuthenticateUser
-{
-    public class AuthenticateUserValidator : AbstractValidator<AuthenticateUserCommand>
-    {
-        public AuthenticateUserValidator()
-        {
-            RuleFor(x => x.Email)
-                .NotEmpty()
-                .EmailAddress();
-
-            RuleFor(x => x.Password)
-                .NotEmpty()
-                .MinimumLength(6);
-        }
-    }
-}
