+++ src/Ambev.DeveloperEvaluation.Application/Auth/Commands/AuthenticateUser/AuthenticateUserResult.cs
@@ -1,15 +1,10 @@
-namespace Ambev.DeveloperEvaluation.Application.Auth.AuthenticateUser;
+namespace Ambev.DeveloperEvaluation.Application.Auth.Commands.AuthenticateUser;
 
 /// <summary>
 /// Represents the response after authenticating a user
 /// </summary>
 public sealed class AuthenticateUserResult
 {
-    /// <summary>
-    /// Gets or sets the authentication token
-    /// </summary>
-    public string Token { get; set; } = string.Empty;
-
     /// <summary>
     /// Gets or sets the user's unique identifier
     /// </summary>
@@ -29,6 +24,11 @@ public sealed class AuthenticateUserResult
     /// Gets or sets the user's phone number
     /// </summary>
     public string Phone { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Gets or sets the authentication token
+    /// </summary>
+    public string Token { get; set; } = string.Empty;
 
     /// <summary>
     /// Gets or sets the user's role
