+++ src/Ambev.DeveloperEvaluation.Application/Auth/Commands/AuthenticateUser/AuthenticateUserValidator.cs
@@ -1,0 +1,17 @@
+using FluentValidation;
+
+namespace Ambev.DeveloperEvaluation.Application.Auth.Commands.AuthenticateUser;
+
+public class AuthenticateUserValidator : AbstractValidator<AuthenticateUserCommand>
+{
+    public AuthenticateUserValidator()
+    {
+        RuleFor(x => x.Email)
+            .NotEmpty()
+            .EmailAddress();
+
+        RuleFor(x => x.Password)
+            .NotEmpty()
+            .MinimumLength(6);
+    }
+}
